requirements.txt
selenium
Pillow
pytesseract
python-dotenv == 0.13.0